+++ src/client/items/history.ts
@@ -8,6 +8,8 @@
 import * as fs from 'fs';
 import { homedir } from 'os';
 import { Event, EventEmitter, MarkdownString, TreeDataProvider, TreeItem, TreeItemCollapsibleState } from 'vscode';
+import path = require('path');
+import moment = require('moment');
 
 const cfgDir = homedir() + '/.vscode/';
 const historyPath = cfgDir + 'q-query-history.json';
@@ -17,6 +19,7 @@ type History = {
     time: Date,
     duration: number,
     query: string,
+    errorMsg: string,
 }
 
 export default class HistoryTreeItem extends TreeItem
@@ -29,27 +32,39 @@ export default class HistoryTreeItem extends TreeItem
     duration: number;
     _parent: TreeItem | null;
     _children: TreeItem[] = [];
+    errorMsg = '';
+    contextValue = 'history';
     public static currentHistoryTree: HistoryTreeItem;
 
     public static createHistoryTree(): HistoryTreeItem {
         if (!HistoryTreeItem.currentHistoryTree)
-            HistoryTreeItem.currentHistoryTree = new HistoryTreeItem({ uniqLabel: 'root', time: new Date(), duration: 0, query: '' }, null);
+            HistoryTreeItem.currentHistoryTree = new HistoryTreeItem({ uniqLabel: 'root', time: new Date(), duration: 0, query: '', errorMsg: '' }, null);
         return HistoryTreeItem.currentHistoryTree;
     }
 
     private constructor(history: History, parent: TreeItem | null) {
-        super(history.uniqLabel.replace(',', '-') + ' | ' + history.time.toISOString(), TreeItemCollapsibleState.None);
+        super(history.uniqLabel.replace(',', '-') + ' | ' + moment(history.time).format('HH:mm:ss'), TreeItemCollapsibleState.None);
         this.uniqLabel = history.uniqLabel;
         this.query = history.query;
         this.time = history.time;
         this.duration = history.duration;
         this._parent = parent;
+        this.errorMsg = history.errorMsg;
         const mdString = new MarkdownString();
-        mdString.appendMarkdown(
-            `- server: ${history.uniqLabel.replace(',', '-')}\n- time: ${history.time.toISOString()}\n- duration: ${history.duration}\n- query:`);
-        mdString.appendCodeblock(history.query, 'q');
+        mdString.appendMarkdown(`- server: ${history.uniqLabel.replace(',', '-')}\n`);
+        mdString.appendMarkdown(`- time: ${moment(history.time).format('YYYY.MM.DD HH:mm:ss.SSS')}\n`);
+        mdString.appendMarkdown(`- duration: ${history.duration}\n`);
+
+        if (this.errorMsg) {
+            mdString.appendMarkdown('- error:');
+            mdString.appendCodeblock(this.errorMsg, 'q');
+        } else {
+            mdString.appendMarkdown('- query:');
+            mdString.appendCodeblock(history.query, 'q');
+        }
         this.tooltip = mdString;
     }
+
     refresh(): void {
         if (this._parent) {
             return;
@@ -69,6 +84,10 @@ export default class HistoryTreeItem extends TreeItem
             }
             fs.writeFileSync(historyPath, '[]', 'utf8');
         }
+    }
+
+    public static dumpHistory(): void {
+        fs.writeFileSync(historyPath, JSON.stringify(HistoryTreeItem.currentHistoryTree._children, null, 4), 'utf8');
     }
 
     getParent(): TreeItem | null {
@@ -93,4 +112,21 @@ export default class HistoryTreeItem extends TreeItem
         HistoryTreeItem.currentHistoryTree._children.unshift(new HistoryTreeItem(history, HistoryTreeItem.currentHistoryTree));
         HistoryTreeItem.currentHistoryTree.refresh();
     }
+
+    // @ts-ignore
+    get iconPath(): { light: string, dark: string } {
+        if (this.errorMsg) {
+            return {
+                light: path.join(__filename, '../../assets/svg/light/error.svg'),
+                dark: path.join(__filename, '../../assets/svg/dark/error.svg')
+            };
+        } else {
+            return {
+                light: path.join(__filename, '../../assets/svg/light/pass.svg'),
+                dark: path.join(__filename, '../../assets/svg/dark/pass.svg')
+            };
+        }
+    }
+
+
 }